requirements.txt
playwright==1.39.0
python-dotenv==1.0.0
schedule==1.2.1
loguru==0.7.2
PyQt6==6.5.2 